requirements.txt
fastapi
pydantic
openai
psycopg2-binary